+++ packages/chatbot/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ownxai/chatbot
 
+## 0.1.2
+
+### Patch Changes
+
+- @ownxai/db@0.1.2
+- @ownxai/sdk@0.1.5
+- @ownxai/tools@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes