+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ownxai/sdk
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @ownxai/shared@0.1.2
+
+## 0.1.4
+
+### Patch Changes
+
+- Update @ownxai/sdk
+
 ## 0.1.3
 
 ### Patch Changes