+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ownxai/shared
 
+## 0.1.2
+
+### Patch Changes
+
+- Update @ownxai/shared
+
 ## 0.1.1
 
 ### Patch Changes